(hel-342) replacer les view model aux bons endroits
